--- scripts/bundles/auto-prefixer.js
@@ -1,2 +1,0 @@
-exports.postcss = require('postcss');
-exports.autoprefixer = require('autoprefixer');
